+++ simple-jwt-login/autoload.php
@@ -1,6 +1,7 @@
 <?php
 
 if (!defined('ABSPATH')) {
+    /** @phpstan-ignore-next-line  */
     exit;
 } // Exit if accessed directly