+++ simple-jwt-login/simple-jwt-login.php
@@ -13,6 +13,7 @@
 use SimpleJWTLogin\Modules\SimpleJWTLoginSettings;
 
 if (! defined('ABSPATH')) {
+    /** @phpstan-ignore-next-line  */
     exit;
 } // Exit if accessed directly