+++ Changelog.md
@@ -4,6 +4,7 @@
 - Add filter to allow the change for authentication payload
 - Change how we log in the user on while using the "protect endpoint" feature
 - Refactor Route Service getUserFromJWT method
+- Update License to GPL v3
 
 ## 3.4.7 (02 October 2022)
 - Remove code vulnerability from the JWT library