+++ Changelog.md
@@ -7,7 +7,7 @@
 - Update License to GPL v3
 - Sanitize data from request
 - Fix password that contains special characters [#50](https://github.com/nicumicle/simple-jwt-login/issues/50)
-- `/auth/validate` endpoint support both `GET` and `POST` methods 
+- `/auth/validate` endpoint supports both `GET` and `POST` methods 
 
 ## 3.4.7 (02 October 2022)
 - Remove code vulnerability from the JWT library